containers: move to new self-contained setup
Move the container definitions into ./lib/containers/ and keep
everything they need in that sub-directory. This allows separating the
entire docker-context from everything else in the repository.

Furthermore, adapt the CI to avoid the makefile, since the docker-buildx
commands are stable now and can be relied upon in our workflows.

Signed-off-by: David Rheinsberg <david@readahead.eu>